String-utf8 problem fixed
