+++ clar2wasm/src/tools.rs
@@ -3,7 +3,7 @@
 //! in production. The `tools` module is only available when the
 //! `developer-mode` feature is enabled.
 
-use std::collections::{BTreeMap, HashMap};
+use std::collections::HashMap;
 
 use clarity::consts::CHAIN_ID_TESTNET;
 use clarity::types::StacksEpochId;
@@ -13,10 +13,7 @@ use clarity::vm::contracts::Contract;
 use clarity::vm::costs::LimitedCostTracker;
 use clarity::vm::database::ClarityDatabase;
 use clarity::vm::errors::{Error, WasmError};
-use clarity::vm::types::{
-    CharType, ListData, OptionalData, PrincipalData, QualifiedContractIdentifier, ResponseData,
-    SequenceData, StandardPrincipalData, TupleData, UTF8Data,
-};
+use clarity::vm::types::{PrincipalData, QualifiedContractIdentifier, StandardPrincipalData};
 use clarity::vm::{execute_v2 as execute, ClarityVersion, ContractContext, Value};
 
 use crate::compile;
@@ -181,108 +178,6 @@ pub fn evaluate_at(
 pub fn evaluate(snippet: &str) -> Result<Option<Value>, ()> {
     evaluate_at(snippet, StacksEpochId::latest(), ClarityVersion::latest()).map_err(|_| ())
 }
-pub fn unicode_to_byte_sequence(input: Value) -> Value {
-    match &input {
-        Value::Sequence(SequenceData::List(list_data)) => {
-            let mut processed_data = Vec::new();
-
-            for val in &list_data.data {
-                processed_data.push(unicode_to_byte_sequence(val.clone()));
-            }
-
-            Value::Sequence(SequenceData::List(ListData {
-                data: processed_data,
-                type_signature: list_data.type_signature.clone(),
-            }))
-        }
-        Value::Response(response_data) => Value::Response(ResponseData {
-            committed: response_data.committed,
-            data: Box::new(unicode_to_byte_sequence(*(response_data.data).clone())),
-        }),
-        Value::Optional(optional_data) => {
-            if let Some(data) = &optional_data.data {
-                Value::Optional(OptionalData {
-                    data: Some(Box::new(unicode_to_byte_sequence(*data.clone()))),
-                })
-            } else {
-                Value::Optional(optional_data.clone())
-            }
-        }
-        Value::Tuple(tuple_data) => {
-            let mut btree_map = BTreeMap::new();
-            for (name, value) in tuple_data.data_map.iter() {
-                btree_map.insert(name.clone(), unicode_to_byte_sequence(value.clone()));
-            }
-            Value::Tuple(TupleData {
-                data_map: btree_map,
-                type_signature: tuple_data.type_signature.clone(),
-            })
-        }
-        Value::Sequence(SequenceData::String(CharType::UTF8(u))) => {
-            println!("Input Data :{:?}", u.data);
-
-            let scalar_bytes = u.data.clone();
-
-            let mut utf8_bytes: Vec<Vec<u8>> = Vec::new();
-
-            for bytes in scalar_bytes {
-                // Reconstruct the scalar value assuming big-endian byte order
-                let mut scalar_value: u32 = 0;
-                let num_bytes = bytes.len();
-                for (i, &byte) in bytes.iter().enumerate() {
-                    scalar_value |= (byte as u32) << (8 * (num_bytes - 1 - i));
-                }
-                println!("Scalar value: {scalar_value}");
-                // Convert the scalar value to a char if it's a valid Unicode scalar value
-                if let Some(character) = std::char::from_u32(scalar_value) {
-                    if character.is_ascii() {
-                        utf8_bytes.push(vec![character as u8]);
-                    } else {
-                        let mut buf = [0; 4];
-                        let encoded_str = character.encode_utf8(&mut buf);
-                        utf8_bytes.push(encoded_str.as_bytes().to_vec());
-                    }
-                }
-            }
-
-            println!("Output UTF-8 Data: {:?}", utf8_bytes);
-
-            Value::Sequence(SequenceData::String(CharType::UTF8(UTF8Data {
-                data: utf8_bytes,
-            })))
-        }
-        _ => input,
-    }
-}
-
-pub fn unicode_scalars_from_string(str: String) -> Value {
-    let validated_utf8_str = std::str::from_utf8(str.as_bytes()).unwrap();
-    let mut data = vec![];
-
-    for char in validated_utf8_str.chars() {
-        let scalar_value = char as u32; 
-        let mut scalar_bytes = scalar_value.to_be_bytes().to_vec();
-        scalar_bytes.retain(|&x| x != 0); 
-
-        data.push(scalar_bytes);
-    }
-
-    
-    Value::Sequence(SequenceData::String(CharType::UTF8(UTF8Data { data })))
-}
-
-#[allow(clippy::result_unit_err)]
-pub fn parse_utf8(input: Result<Option<Value>, ()>) -> Result<Option<Value>, ()> {
-    if let Ok(ok_value) = input {
-        if let Some(v) = ok_value {
-            Ok(Some(unicode_to_byte_sequence(v)))
-        } else {
-            Ok(None)
-        }
-    } else {
-        input
-    }
-}
 
 pub fn crosscheck(snippet: &str, expected: Result<Option<Value>, ()>) {
     let compiled = evaluate_at(snippet, StacksEpochId::latest(), ClarityVersion::latest());
@@ -296,15 +191,10 @@ pub fn crosscheck(snippet: &str, expected: Result<Option<Value>, ()>) {
         &interpreted
     );
 
-    let parsed_expected = parse_utf8(expected.clone());
-
-    println!("Expected Original:{:?}", expected);
-    println!("Expected Parsed:{:?}", parsed_expected);
 
     assert_eq!(
         compiled.map_err(|_| ()),
-        parsed_expected,
-        // expected,
+        expected,
         "Not the expected result"
     );
     println!("-------------------------------------");
